Move the Kubernetes scripts files under Kubernetes folder
