+++ source/Octopus.Tentacle/Kubernetes/Scripts/KubernetesJobScriptExecutor.cs
@@ -4,9 +4,9 @@ using System.Threading.Tasks;
 using Octopus.Diagnostics;
 using Octopus.Tentacle.Configuration.Instances;
 using Octopus.Tentacle.Contracts.ScriptServiceV3Alpha;
-using Octopus.Tentacle.Kubernetes;
+using Octopus.Tentacle.Scripts;
 
-namespace Octopus.Tentacle.Scripts.Kubernetes
+namespace Octopus.Tentacle.Kubernetes.Scripts
 {
     public class KubernetesJobScriptExecutor : IScriptExecutor
     {
@@ -27,7 +27,7 @@ namespace Octopus.Tentacle.Scripts.Kubernetes
 
         public IRunningScript ExecuteOnBackgroundThread(StartScriptCommandV3Alpha command, IScriptWorkspace workspace, ScriptStateStore scriptStateStore, CancellationToken cancellationToken)
         {
-            var runningScript = new RunningKubernetesJobScript(workspace, workspace.CreateLog(), command.ScriptTicket, command.TaskId, cancellationToken, log, scriptStateStore, jobService, containerResolver, appInstanceSelector);
+            var runningScript = new RunningKubernetesJob(workspace, workspace.CreateLog(), command.ScriptTicket, command.TaskId, cancellationToken, log, scriptStateStore, jobService, containerResolver, appInstanceSelector);
 
             Task.Run(async () =>
             {
